refactor: empty the create_api_resources closure (#667)
The last four groups move out: lifecycle (create/renew/reissue/jobs),
certificates (list/detail), deployment, and discovery. With them goes the TLS
probe cluster they shared, into tls_probe.py.

create_api_resources held 42 Resource classes as closures over ten managers.
It now holds none: it builds the groups, binds the two namespaces created here
rather than by factory.py, and returns the mapping.

  complexity   559 -> 1
  resources.py 4040 -> 230 lines

The manager aliases and helper wrappers went with the last class — flake8
reporting all ten as unused was the signal that nothing in the function used
them any more.

Five tests had to be repointed, and each one was checked to still bite rather
than merely to pass:

Two source-reading ratchets asserted on strings inside
create_api_resources. One looked for `certificate_manager.domain_lock(domain)`,
which no longer exists in that form — the extraction renamed it to
`ctx.certificates`. Both now read resources_certificates and are
mutation-verified.

test_deployment_status_route and test_deploy_status_cache_invalidation stub the
TLS probe. A stub must be installed in the module that CALLS it: resources.py
still re-exports these helpers, so patching that copy would have left the call
site untouched and the tests would have passed while stubbing nothing. Verified
by pointing them back at resources.py — eight of fourteen fail, so the stubs
are genuinely in effect.

test_audit_hardening patched `modules.api.resources.Path.resolve`; resources.py
no longer imports Path.

The new guard states the property #667 was opened for, rather than a line
count: no Resource class may live inside that closure. Each one there is an
endpoint that cannot be built, or tested, without the whole manager graph —
which is what #662 needs.
